+++ web/Controllers/HomeController.cs
@@ -25,6 +25,7 @@
 
 
 using System.Web.Mvc;
+using ASC.Api.Web.Help.Helpers;
 
 namespace ASC.Api.Web.Help.Controllers
 {
@@ -33,6 +34,12 @@ namespace ASC.Api.Web.Help.Controllers
         public ActionResult Index()
         {
             return View();
+        }
+
+        [ValidateInput(false)]
+        public ActionResult Search(string query)
+        {
+            return View(GCustomSearch.Search(query));
         }
     }
 }